TCR-27 ui-service-interactions
This module DOES NOT meet the following module acceptance criteria:

* Have WCAG 2.1 AA compliance as measured by a current major version of axe DevTools Chrome Extension
* Follow relevant existing UI layouts, patterns and norms

Additionally, WRT "Module only uses FOLIO interfaces already provided by
previously accepted modules", note that `servint` appears to have been
introduced in early 2021, before the module acceptance process was in
place. This feels more like "this is an exception" rather than "this is
acceptable".

TCR process improvements:
* I am increasingly strongly uncomfortable evaluating license compatibility. I suggest we change the line
> Third party dependencies use an Apache 2.0 compatible license
to
> Includes a report of the licenses used by third-party dependencies
and we can delegate evaluation of that list to a person/body with appropriate credentials for this kind of thing. IOW, IANAL and I really really don't want to be responsible for making definitive statements about license compatibility. Example tools in NPM-land:
  * apache2-license-checker
  * license-checker
* Do we/should we have a Source Of Record for application documentation? Is it OK to link to the wiki or should apps point to https://docs.folio.org/?

Refs TCR-27