Expand criterium for consumed APIs; mention requires/okapiInterfaces/...
This is the current acceptance criterium:

-* Module descriptor MUST include interface requirements for all consumed APIs (3, 5)
-  * _This is not applicable to libraries_

This should be more actionable.

Evaluators need to browse through the source code to find out the consumed APIs and then check that the API is listed.
Therefore the parts of the sentence should be swapped to start with "For each consumed API".

Evaluators of backend modules need to look up the interface requirement in the `"requires"` or `"optional"` section
of the module descriptor. The section names should be mentioned in the criterium to make it more actionable. The name
references the relevant part of the module descriptor documentation making the criterium easier to understand (the
preceding criterium links to the module descriptor documentation).
The criterium about the provided interfaces already mentions the `"provides"` section, similarly the the
`"requires"` and `"optional"` sections should be mentioned.

Currently the criterium about the interface requirements in the module descriptor also applies to frontend modules.
This can be simplified because the build process generates the module descriptor from the package.json file, and
there's already a criterium that it "MUST produce a valid module descriptor".

Therefore we can change the criterium for the frontend to require the interfaces of the consumed APIs to
be in the `"okapiInterfaces"` or `"optionalOkapiInterfaces"` section of the package.json file. This is more actionable
because the package.json file can be edited while the module descriptor doesn't exist in the code repository.
A note should link to the package.json documentation:
https://github.com/folio-org/stripes/blob/master/doc/dev-guide.md#the-package-file-stripes-entry

Frontend shared libraries should also contain complete `"okapiInterfaces"` or `"optionalOkapiInterfaces"` sections.
The current criterium doesn't require this because a frontend shared library doesn't generate a module descriptor.
For a complete okapi interface dependency documentation this change adds the requirement for frontend shared libraries.

The common criterium about interface requirements is split into frontend and backend criterium and moved from
the Shared/Common section to the Frontend section and the Backend section.